refactor(route/jike): rename `common` -> `utils`
