+++ lib/v2/jike/topic.js
@@ -1,5 +1,5 @@
 const got = require('@/utils/got');
-const common = require('./common');
+const { emptyResponseCheck, topicDataHanding } = require('./utils');
 const cheerio = require('cheerio');
 const dayjs = require('dayjs');
 
@@ -13,7 +13,7 @@ module.exports = async (ctx) => {
     const raw = $('[type = "application/json"]').html();
     const data = JSON.parse(raw).props.pageProps;
 
-    if (common.emptyResponseCheck(ctx, data)) {
+    if (emptyResponseCheck(ctx, data)) {
         return;
     }
 
@@ -23,7 +23,7 @@ module.exports = async (ctx) => {
         link: `https://m.okjike.com/topics/${id}`,
         description: topic.briefIntro,
         image: topic.squarePicture.picUrl || topic.squarePicture.middlePicUrl || topic.squarePicture.thumbnailUrl,
-        item: common.topicDataHanding(data),
+        item: topicDataHanding(data),
     };
     if (id === '553870e8e4b0cafb0a1bef68' || id === '55963702e4b0d84d2c30ce6f') {
         const promises = ctx.state.data.item.map((one) => {
