+++ lib/v2/jike/topicText.js
@@ -1,6 +1,6 @@
 const got = require('@/utils/got');
 const dayjs = require('dayjs');
-const common = require('./common');
+const { emptyResponseCheck } = require('./utils');
 const cheerio = require('cheerio');
 
 module.exports = async (ctx) => {
@@ -13,7 +13,7 @@ module.exports = async (ctx) => {
     const raw = $('[type = "application/json"]').html();
     const data = JSON.parse(raw).props.pageProps;
 
-    if (common.emptyResponseCheck(ctx, data)) {
+    if (emptyResponseCheck(ctx, data)) {
         return;
     }
